+++ CHANGELOG.md
@@ -3,6 +3,23 @@ aws-parallelcluster-cookbook CHANGELOG
 
 This file is used to list changes made in each version of the AWS ParallelCluster cookbook.
 
+3.9.0
+------
+
+**ENHANCEMENTS**
+- Add support for installing Intel OneAPI Base Toolkit and HPC Toolkit, and Intel Python.
+  - Intel OneAPI Base Toolkits: 2023.2.0
+  - Intel OneAPI HPC Toolkits: 2023.2.0
+  - Intel Python: 2023.2.0
+  - Critical Update for Intel oneAPI DPC++/C++ Compiler: 2023.2.1
+  - Critical Update for Intel Fortran Compiler & Intel Fortran Compiler Classic: 2023.2.1
+
+**CHANGES**
+- Upgrade third-party cookbook dependencies:
+  - nfs-5.1.2 (from nfs-5.0.0)
+
+**BUG FIXES**
+
 3.8.0
 ------
 
@@ -28,6 +45,7 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 
 **BUG FIXES**
 - Fix inconsistent scaling configuration after cluster update rollback when modifying the list of instance types declared in the Compute Resources.
+- Fix users SSH keys generation when switching users without root privilege in clusters integrated with an external LDAP server through cluster configuration files.
 
 3.7.2
 ------